Continuation of #4524.

RELEASE NOTES:

client: improve performance when multiple interceptors are used