@zasweq 2021-06-30
Can you please provide some benchmarks for this change for the Client Side interceptors?