@amenzhinsky 2021-07-01
Added benchmaks and here's the result:

```
benchmark                                        old ns/op     new ns/op     delta
BenchmarkChainClientUnaryInterceptors/1-12       7.68          7.69          +0.10%
BenchmarkChainClientUnaryInterceptors/3-12       146           151           +3.77%
BenchmarkChainClientUnaryInterceptors/5-12       318           182           -42.69%
BenchmarkChainClientUnaryInterceptors/10-12      767           222           -71.07%
BenchmarkChainClientStreamInterceptors/1-12      7.02          6.95          -1.07%
BenchmarkChainClientStreamInterceptors/3-12      133           164           +23.44%
BenchmarkChainClientStreamInterceptors/5-12      276           180           -34.84%
BenchmarkChainClientStreamInterceptors/10-12     551           209           -62.16%

benchmark                                        old allocs     new allocs     delta
BenchmarkChainClientUnaryInterceptors/1-12       0              0              +0.00%
BenchmarkChainClientUnaryInterceptors/3-12       2              3              +50.00%
BenchmarkChainClientUnaryInterceptors/5-12       4              3              -25.00%
BenchmarkChainClientUnaryInterceptors/10-12      9              3              -66.67%
BenchmarkChainClientStreamInterceptors/1-12      0              0              +0.00%
BenchmarkChainClientStreamInterceptors/3-12      2              3              +50.00%
BenchmarkChainClientStreamInterceptors/5-12      4              3              -25.00%
BenchmarkChainClientStreamInterceptors/10-12     9              3              -66.67%

benchmark                                        old bytes     new bytes     delta
BenchmarkChainClientUnaryInterceptors/1-12       0             0             +0.00%
BenchmarkChainClientUnaryInterceptors/3-12       96            80            -16.67%
BenchmarkChainClientUnaryInterceptors/5-12       192           80            -58.33%
BenchmarkChainClientUnaryInterceptors/10-12      432           80            -81.48%
BenchmarkChainClientStreamInterceptors/1-12      0             0             +0.00%
BenchmarkChainClientStreamInterceptors/3-12      96            80            -16.67%
BenchmarkChainClientStreamInterceptors/5-12      192           80            -58.33%
BenchmarkChainClientStreamInterceptors/10-12     432           80            -81.48%
```