@zasweq 2021-07-02
Hey, the benchmarks show a positive increase. Sorry, but the team can't take the PR with all the positive increases. The Server Side one was fantastic though :).